This has two things:    1) musicomp to abc music notation converter. (Not done yet, see NOTDONEYET)    2) A calculator using the precedence that the Plato system used for calc       per arrow / compute execution, not programs. (i.e. implied multiply :) )
The musicomp format was from 1973/1974/1975/1976 era of the Plato system at
the University of Illinois.

It started out as a single square wave ability. [Thus the single voice songs.]
It then changed to a four voice saw tooth format, with 7 levels volume level,
although only 5,6,7 (of 0 through 7) were able to be heard. *sigh*

The third version was multi-voice with loadable wave forms. The cyber1 system
does NOT have any of the music related to that.  Too bad, blue danube and
beethoven's fifth symphony were entered to provide an example of "long" tunes.

Now ... abc format remembers a sharp/flat/natural changes within a measure/bar.
The musicomp format did not. There are WARNING messages given by musicomp2abc
when such an event is noted. Sometimes they are 'mistakes', and sometimes they
need to have the sharp/natural/flat removed/inverted.

Several commands do not make sense. The "key" exists. natural/flat/sharp is not
available in the abc format. Fortunately, most of the songs didn't use that, or
were easily converted to use "key". :)

There are only four files that used the "macro", "putd", etc. commands. They
are not converted.

I like "make" and Makefiles. Comment in the "# SONG += Coke" (as example) to
have it converted to abc format, and abc2midi to create playable (via timidity
for example), and yaps (abc format to postscript) for display.

Several extra commands were added to the old musicomp format:
    meter    5/4
    bars     3				# number of bars printed / score line.
    clef     bass,treble,treble,treble  # Which clef to print per voice.
    voice    4    66,57,56,48		# Midi instrument to use for the voice.

It is suggested that if number of voices change within the program -- usually
"voice 1" and "voice 4", that "cis 4" is used instead of "voice 1" so that the
instrument stays the same for the single voice notes, and the last voice of
the four voice music.  (lowest note is commonly first, and highest last)
Intermediate 2 and 3 are best to have the extra commas added to make it four.
------------------------------------------------------------------------------
------------------------------------------------------------------------------