Add --abc printing of bar lines joined together in vertical mode.     This is a note following a comma, with a space. 	"v4: 3c8, d,e,f"     This would have the bar lines for d,e,f joined together, and 'c' is     standing alone.
Add legato printing in --abc format. (gliding/legato/slurred notes)
    Example: "v4: 3c8l,dl,e, f"
    Would have the c,d,e bar lines connected, and a glide/slur/legato
    between the c,d,e with a glide/legato/slur between them

Not done yet: Legato playing.
Research implies that the attack portion of the note always plays in
fluidsynth and timidity and musescore and ... .
It appears that legato (midi control 68) does nothing in this programs.
The sound fonts do not have it defined either.
Marshall (me, myself, and I) is currently thinking that changing fluidsynth
to skip the Attack portion of a note and go immediately to the sustain section
may be an easy attempt at legato. No idea how/if it would work, or if it would
sound reasonable.