A case where "," and " " (space) are the same, but one MAY NOT expect it. The ties went to the auto-added rests for the timpani.
    measure 109
    timpani: r2d
    measure 110
    timpani: vol(ff) t(1a,1a,12) 2d

The trill is of length "2d" (dotted half note - remembered from last
note / rest length), and then a note '2d' with remembered length of '2d' too.

This then means that all voices in the measure will be filled with 'rests' so
that they are up to length (2d + 2d) or (1d - whole note dotted).